Add configs folder
